clean up client test
