+++ instrumentation/finagle-http-23.11/javaagent/src/test/java/io/opentelemetry/javaagent/instrumentation/finagle/v23_11/ServerH2Test.java
@@ -6,8 +6,6 @@
 package io.opentelemetry.javaagent.instrumentation.finagle.v23_11;
 
 import static io.opentelemetry.instrumentation.netty.v4_1.internal.ProtocolSpecificEvent.SWITCHING_PROTOCOLS;
-import static io.opentelemetry.instrumentation.testing.junit.http.AbstractHttpClientTest.CONNECTION_TIMEOUT;
-import static io.opentelemetry.instrumentation.testing.junit.http.AbstractHttpClientTest.READ_TIMEOUT;
 import static io.opentelemetry.instrumentation.testing.junit.http.ServerEndpoint.SUCCESS;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
@@ -17,7 +15,6 @@ import com.twitter.finagle.Service;
 import com.twitter.finagle.http.Request;
 import com.twitter.finagle.http.Response;
 import com.twitter.finagle.http2.param.PriorKnowledge;
-import com.twitter.finagle.service.RetryBudget;
 import com.twitter.util.Await;
 import com.twitter.util.Duration;
 import io.opentelemetry.api.common.Attributes;
@@ -63,15 +60,9 @@ class ServerH2Test extends AbstractServerTest {
   void h2ProtocolUpgrade() throws Exception {
     URI uri = URI.create("http://localhost:" + port + SUCCESS.getPath());
     Service<Request, Response> client =
-        Http.client()
+        Utils.createClient(Utils.ClientType.DEFAULT)
+            // must use http2 here
             .withHttp2()
-            .withTransport()
-            .readTimeout(Duration.fromMilliseconds(READ_TIMEOUT.toMillis()))
-            .withTransport()
-            .connectTimeout(Duration.fromMilliseconds(CONNECTION_TIMEOUT.toMillis()))
-            // disable automatic retries -- retries will result in under-counting traces in the
-            // tests
-            .withRetryBudget(RetryBudget.Empty())
             .newService(uri.getHost() + ":" + uri.getPort());
 
     Response response =
